Refactoring Exif code
